Use marker clusterer when zoom is below minimal level
Instead of fetching no markers, markers are fetched with the is_thin option, which means they don't have any field except for the essential ones (id, location). That is enough for showing the clusters and otherwise the performance would suffer.
When zooming back in the caches are cleared and the markers are re-fetched with all their fields again.