Fix #189: in cluster mode, filter on server and send real thin
Do not perform filtering on client when zoom level is below the minimal zoom (cluster mode). This is so that we can send the markers from the server with only the id, longitude and latitude fields. The client does not need to know these fields because it does not use them for filtering, and shows clusters rather than individual markers.

When we are in cluster mode, we need to reset all the markers whenever anything changes, because we do not filter locally.