Change center_changed back to mouseup
We don't want the markers to load each time the map is moved because it causes a performance decrease.
This opens back #175 but it's OK because the accidents would be in the background anyway and the user can close the infoWindow to fetch them.

Also fix the first_name problem in marker.