Partial fix for #248: create discussion marker when posting to Disqus
The thread is now tied to the marker location (lat & lng), and a marker
is created when posting a comment.
However, the marker is not yet sent to the server, and it will disappear
when reloading the app. It will not show for other users either.
Still need to send an Ajax call to create the marker in the DB and show
it with all the accident markers on the map.