If failed loading years, keep trying
Too many connections may cause an OperationalError when trying to read
the years in order to create the file 'years.js'.
In that case, any subsequent request will result in an error, and the
dyno has to be restarted to fix it. Instead, we'll now retry creating
the file until successful.