Use ESLint to lint the JavaScript code
This patch adds an ESLint configuration, together with the necessary fixes, which are mostly adding
a missing semicolon. ESLint will run in Travis from now as well.

The reason why we turned off the `undef` and `unused-vars` errors is because currently the
JavaScript files rely on the fact that they are loaded in the HTML file in the correct order, making
it impossible for ESLint to figure out which variables are defined in other files. Once we find a
better import mechanism we should turn these settings back on.