Parameterized markers filter testing
This patch removes all the hard-coded cases for markers filtering and replace them with a
parameterized testing.

Pytest expands test_markers into multiple tests, containing a test for the entire Cartesian product
of the filtering flags.

Each test checks that results that were supposed to be filtered out got in, and at the end of all
tests the marker_counter fixtures makes sure that we got the expected amount of result in total by
comparing it against a hard-coded number. This number should be updated if we ever add more
parameters to the test, or import additional marker data into the testing environment.