Allow scrolling info windows by using the mouse wheel (#747)
Currently, trying to use the mouse wheel to scroll an info window results in the map zooming in or
out instead of scrolling the info window. This patch fixes this behavior by settings the map
gesture handling to cooperative only when an info window is active. This means that while an info
window is active, zooming is done by holding the control key together with the mouse wheel, allowing
to use the mouse wheel to scroll the info window. Once the info window is closed, zooming is
performed only by using the mouse wheel.