Unify the database initialization (fix #752)
So far we've had two methods for initializing the database - one is using flask-sqlalchemy and the
other is using the session object in the database module.

When specifying a relative path for an SQLite database, the default behavior of SQLAlchemy is to
use the file relative to the current working directory. However, the behavior of flask-sqlalchemy
is different. flask-sqlalchemy tries to use the file relative to the Flask application root.

Before 0274aeabd17cfcefb218656e029ce5927cda91bb the working directory and the Flask application root
were effectively the same directory, but since now they are different directories, it causes two
different SQLite databases to be created - one in the working directory and the other in the Flask
application root.

The solution is to unify these methods and always use flask-sqlalchemy to initialize the database.

The use of SQLite database will also be checked in Travis from now on.