Reorganize the use of Alembic (fix #757)
This patch reorganizes the use of Alembic.

Our current revisions were out of sync and thus irrelevant, so they were removed.

An initial version was made which now contains the instructions to create the entire schema.

The init_db subcommand was removed as users should now use the alembic command to initialize the
schema. Both the README and the CI were updated accordingly.

Contributors that make pull requests which alter the schema should prepare the appropriate alembic
version and commit it with their patch