UI Should not pass the age group parameter if all groups are selected
Prior to this patch, the first request that was made by Anyway to the server did not specify the
age_group parameter.

When the user clicked on one of the filters, the age group filters were evaluated. If the user did
no click on an age group button, the age_group parameter would contain all
groups ("1,2,...,18"). However, this is not same as not specifying the age_group parameters at
all. Despite the fact that all age groups were specified, markers from United do not contain age
groups, and thus were not selected.

The proper solution is to specify an empty age_group parameter when all age groups are selected.