Properly configure the age group relationship
Since a proper relationship wasn't defined so far, we used the join method in order to filter by age
groups of the people involved in an accident.

The problem was that the `count` method that we used to report the total number of records counted
the number of rows in the query, which is greater than the number of accidents objects due to
joining the involved people's records.

Once the relationship is properly configured, we can filter the table by using the "any" method,
which makes `count` return the correct result.