Add widget (Issue #1715)
Widget that returns the number seriously injured or killed in accidents with bicycles, electric bicycles, or scooters per year in a given city.
It is an urban widget as it is only relevant when the newsflash takes place in a city (and not between them) and is included when there have been relevant accidents in that city in the last year. (Ideally in the future - the api of the newsflash would include the vehicles involved, and then it could be included when the newsflash contains one of these vehicles instead).
It returns the seriously injured and fatalities per year per vehicle, plus all three together per year.